Сделано дз по модульности
