+++ js/get-random-integer.js
@@ -1,0 +1,9 @@
+const getRandomInteger = (min, max) => { // функция генерации случайного целого числа из диапазона
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+export {getRandomInteger};
